[SW-103389] habanalabs/gaudi3: add Fabric Serialization Enhancement support
Due to Synopsis's PCIe controller limitation of the number of inbound
write access to 128, while inflight (outstanding) transactions number
is 256 - we must use early write response + fabric serialization.

1. With 'Early Write Response', the ASIC's mesh immediately reports a
bresp for either a LBW/HBW write, thus allows taking advantage of the
256 possible inflights.
2. 'Fabric Serialization', is a feature that has existed in all of our
previous ASICs yet was never used. With it in Gaudi3, a LBW (HBW) write
will not be sent until all previous HBW (LBW) writes were acknowledged
by the mesh, and vice versa.

However, after careful review with the SW, we’ve learned that it might
cause a big performance impact.
The reason is that we always have LBW access towards the device
(driver --> NIC). This will cause bubbles between the HBW & LBW accesses.
Because, currently, there's only a black or white serialization condition-
any LBW access will wait to flush all HBW traffic.

Instead, we enhance this feature by monitoring only specific address
ranges. e.g., the only LBW accesses that will wait for the flush of the HBW
traffic, would be the ones that belong to any of the configured ranges.
That way, we can distinguish between different types of LBW writes.

Later on, this will be wrapped for the user's own use (where the user will
simply issue a write to a specific DUMMY register to get all HBW traffic
flushed).

Note that this feature is currently disabled by default.

Change-Id: If168aa059686063832bad28ce3267de83d0c4617
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>