[SW-99610] habanalabs: support scale-out collective connections
So far, every request for a new collective connection was resulted in a
per-port QP allocation with the same QP id, and each QP was of course
needs to be backed-up with WQ memory.
The problem with that approach was that some ports (scale-out) need much
more QPs than the others (scale-up).

This change let the user the option to allocate a different number of
collective QPs for each type of ports, by adding new input parameter
"is_scale_out" to the alloc_coll_conn API.

Note that every type of collective connection needs to be backed-up with
its own WQ array, hence a new WQ array types were added as-well.

Change-Id: Ib3b444c64918436b8a95b9c52314a107b5893635
Signed-off-by: Sagiv Ozeri <sozeri@habana.ai>