[SW-109644] habanalabs/gaudi3: use completions to signal ETR bufs available
Once an AC (Autonomous Controller) buffer is ready, 'ETR buf store'
will fetch and store it, making it ready for consumption.
So far, the user had to poll on relevant debug IOCTL until buffers
were ready, which is inefficient.
To avoid this inefficiency, kernel-completions were added allowing
user threads to be sent to sleep until new buffers are fetched.
Corresponding threads will be woken upon the retrieval of new buffers.

Change-Id: I6e3abbb77142d4dd754034a9a22265e54050a3b8
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>