[SW-113681] gaudi3: fix skip check for gaudi3_sram_init
The condition in gaudi3_sram_init() to skip the configuration should be
if cache enabled OR in compute reset, and not AND in compute reset.
Moreover, we usually don't use the in compute reset flag to skip
unneeded configurations, so replace it with a dedicated CAP bit.

In addition, remove the comment there as it is inaccurate, because the
cache mode is set when loading the driver and it is not modified when
resetting.

Change-Id: Ic796382fe1e278bbba0bb01fb804370fa6b989a4
Signed-off-by: Tomer Tayar <ttayar@habana.ai>