[SW-0] gaudi2: fix nic init flow
Now that the we get the NIC information from F/W during every reset,
we need to take care of setting the NIC in that stage.
The pre_core_init function will only take care of the cases when F/W is
not running or when we need to ignore the information from it.

This commit also fixes the transition from PAM4 to NRZ mode.

Change-Id: I709c311b906369c27c270ee257acf2c507bf33ab
Signed-off-by: Sagiv Ozeri <sozeri@habana.ai>