[SW-0] nic: fix collective QP destroy flow
Before going to destroy a collective QP, we need to make sure that all
the standard QPs under it were destroyed.
For QPs that weren't asked to be destroyed yet, instead of just free
them, call the QP release function like we do with all other QPs.
In that way, all the QPs will go through the same release flow.

Change-Id: Ic7cd0b79f6b398661249d61df3f813e1dad29dac
Signed-off-by: Sagiv Ozeri <sozeri@habana.ai>