[SW-111592] nic: set QPC_MAX_OTF_PSN_SACK to 1023 to support H9-5457
QPC_MAX_OTF_PSN_SACK holds the number of PSNs that are allowed to be OTF
(On The Fly), when using SACK.

This change causes the TXE to stop sending packets at 1023 OTF rather
then 1024, i.e. avoid a case where the windows is full and the HW bug
may occur.

Change-Id: I15304ba95c7953b50b46fc97a199cd2db9166827
Signed-off-by: Abhilash K V <kvabhilash@habana.ai>
Signed-off-by: David Meriin <dmeriin@habana.ai>