[SW-0] gaudi2/3: fix user_cq memory leak
If the write to the QPC gateway fails, we are relinquishing the
reference to the cq buffer. But again inside the qp_destroy_work, we are
doing a user_cq_put. This is a double put even though we get the
reference only once. Inside the hl_nic_user_cq_put, there is no call to
the user_cq_unset. Inside user_cq_unset only we are destroying the user_cq
and pi memories. Hence there is a memory leak of cq_buffer and cq_pi.

This is fixed by making the qp->res_user_cq/req_user_cq pointer as NULL
once we do a put when QPC gateway write fails. This way, we won't be
calling the 2nd put from qp_destroy_work.

Change-Id: I21715084e45e4f27b70d1d984d413b67b44073bb
Signed-off-by: Ganesh Kumar Gurumurthy <gkgurumurthy@habana.ai>