[SW-113468] habanalabs/gaudi: allow device aquire while in dbg mode
During device aquire, the driver is using a QMAN for clearing some
registers. In order to avoid internal races, the driver verifies
the device is idle before submitting the register clear job.

This check introduces an issue, as debug mode will cause the device
to be non-idle which will lead to device aquire failure.

In order to overcome this issue we can entirely remove the idle
check as the driver is using the QMAN only when there is no active
context.

Change-Id: I891307634dd74bac2ae43619ea20d18ab4f85bfa
Signed-off-by: Ofir Bitton <obitton@habana.ai>