[SW-108259] gaudi3: group all NIC privileged register writes
Privilege register should be configured by FW when it exists.
When there is no FW present, the driver is preforming the configuration.
On coral we run with security assert simulation which the driver enables.
As there is no FW the driver responsible to configure the registers
and for that assertion on coral should be disabled.
NIC blocks were excluded from this security configuration as it was spread
on different functions.
To enable the security configuration, we must group all NIC privilege
registers writing to single place, same as done by compute flow.
This commit addresses this change and enables the security block
configuration.

Change-Id: Ie322a0f8698b5cbf8cd9075abbef290a656457a7
Signed-off-by: Andrey Agranovich <aagranovich@habana.ai>