[SW-116454] habanalabs: consider init_done when poll in sim mode
Up until now, the simulator didn't use the poll macro before the device
was enabled (just before HW init).

Now, introducing the FW simulator, the simulator has to use this macro
in earlier stage (e.g. early init).

In order to be able to poll we need to allow polling when the device is
still disabled but if init sequence is not yet done.

Change-Id: I26c04b37218be65b27315ecafeb879fcf044dc1f
Signed-off-by: Ohad Sharabi <osharabi@habana.ai>