[SW-118219] gaudi3: add skeleton for new EQ handling
Add an initial skeleton for new EQ handling, and move handling of CS
events as a first example for this new path.

Temporarily add an "_old" suffix to the existing gaudi3_handle_eqe()
function.
This function will be removed once the handling of all events is moved
to the new path.

Change-Id: Ie96934907aecdcaeba16937bcce2bab96f85c7b1
Signed-off-by: Tomer Tayar <ttayar@habana.ai>