[SW-0] habanalabs: don't protect privileged PBs with automated security
When LKD is running automated security (i.e., PB configuration to protect
the ASIC of type either secured, privileged, or both), we don't configure
the last 3 PB registers. We do this to be able to reconfigure them upon
a failure followed by a reset procedure.
However, PB reconfiguration appears to be limited only for secured entities
(e.g., LKD) trying to reconfigure privileged PBs.
Hence, for the case of privileged PBs, we don't configure the last 3 PBs.

Change-Id: Ie6d76eb0e5bf2608a08f6baa1956c0cdda5388c7
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>