[SW-119790] habanalabs/gaudi2: increase user interrupt grace time
Currently we support scenarios where a timestamp registration request
of a certain offset is received during the interrupt handling of the
same offset. In this case we give a grace period of up to 100us for
the interrupt handler to finish.
It seems that interrupt handling takes more than expected and should
be optimized. Meanwhile we increase the grace period in order not to
reach timeout which will cause user call to be rejected.

Change-Id: I04aadb58ba849ae3fdf756f9fbe97f90c545aaf0
Signed-off-by: Ofir Bitton <obitton@habana.ai>