[SW-119599] habanalabs: add info when FD released while device still in use
When user closes the device file descriptor, it is checked whether the
device is still in use, and a message is printed if it is.
To make this message more informative, add to this print also the reason
due to which the device is considered as in use.
The possible reasons which are checked for now are active CS and
exported dma-buf.

Change-Id: Iafb7396e30153c64476ed3b1bdfcccc3ee70fcda
Signed-off-by: Tomer Tayar <ttayar@habana.ai>