[SW-0] gaudi2/gaudi3/simulator: avoid wrong error prints during reset
In current implementation, each time simulator is closed while a user
process is still running workloads, we are seeing wrong error prints.
Those errors are true only if running on real asic, but when running
with simulator we should avoid those prints as they are causing a
lot of confusion.

Change-Id: I8ba5e28433de8f37b34b644d5c79d4109338bd58
Signed-off-by: Ofir Bitton <obitton@habana.ai>