[SW-0] gaudi2 phy: reflect the correct link status to the user
When the user wants to get the link status for a port (and in particular
for internal port), he can do that by sending an ioctl.
The status that the LKD will pass to him is according to the pcs_link
field in the common port structure.

So far, in Gaudi2, this field was set in every link status event.
As a result, in case of link toggling that usually happens during PHY
tuning process, this field is set many time as true, but in fact the
link is very unstable and the tuning process can eventually fail and
trigger the PHY reconfigure process.
In that case, the user may query the link status in that time and get
the impression that the link is UP and stable, whereas this is not the
case at all.

This commit fix this potential issue:
In case we got link up event and PHY is not ready yet (the tuning process
hasn't finished), we will just record it in another field.
The actual field that is reflected to the user will be updated only later
when the PHY is ready.

Change-Id: Ic14ece9314340dbd23503dd491de3955c737174a
Signed-off-by: Sagiv Ozeri <sozeri@habana.ai>