[SW-120024] habanalabs: add driver reset IOCTL support for simulator
So far, a failure occurred in a simultaor device (e.g., a page-fault)
was asserted by the simulator device utility, followed by an
aggressive device closure.
It means that, essentially, a simulator couldn't have handled failures.

With this patch, the driver can now execute a reset procedure via an IOCTL,
which the simulator device can use to reset itself rather than simply exit.
If utilized, it will not only prevent an aggressive device closure, but
will allow a quick device recovery.

Change-Id: I1c22b4b59214cdb45177163a41d14083e3d9a4a6
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>