[SW-72317] gaudi3: support cpucp packet send to FW for wqe checker change
WQE checker is dynamically modified upon user call to
set_user_app_params. As this is a privilege register we must
modify it though sending an appropriate ticket to FW.
In gaudi3, WQE checkers are defined per macro and hence we
should pass to FW macro index we need to modify.
Hence, we add to cpucp struct macro_index under the union.

Change-Id: Id7b6ad6536556c108453fabe252d842fa6738d51
Signed-off-by: Andrey Agranovich <aagranovich@habana.ai>