[SW-121332] habanalabs: disable PCI when escalating compute to hard-reset
In case a compute reset has failed or a request for a hard reset has
just arrived, then we escalate current reset procedure from compute
to hard-reset.
In such a case, the FW should be aware of the updated error cause,
and if LKD is the one who performs the reset (rather than the FW),
then we ask the FW to disable PCI access.

As well, we would like to have relevant debug info and therefore
we notify dmesg log of the currently escalating reset type.

Change-Id: If1bd0d323c883a38c9cba53e8259ffb829232af1
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>