[SW-119390] gaudi2/3: refactor nic bringup file separate pldm and fw flows
nic bringup file holds all privilege register accesses. Part of this
configuration should be port into FW, but others are used for PLDM
regardless of FW presence.
For better tracking of code relation, the names were refactored,
and the BFEs that are relevant for debug are set under different
function flow.
Also a check for FW presence was removed as the caller function
guaranties us that we enter this configurations only if FW not present.

Change-Id: I024c59bf84a21a21c9edbe0fc365bf7c4b1e536b
Signed-off-by: Andrey Agranovich <aagranovich@habana.ai>