[SW-120875] gaudi3: stop monitoring for AXI drains through a dedicated work
In Gaudi3, the IRQ aggregator is enabled even when there's no FW,
which means that we can get IRQs when there's no FW.
This is why the AXI drain can be monitored through handle eqe flow.
Unlike in Gaudi2, where we had no IRQ aggregator (and the only way to
receive IRQs was if the FW is enabled) which is why we had to monitor
the AXI drain registers, in Gaudi3 there's no need for such monitoring
(unless interrupt aggregator is disabled).

As well, since reading AXI drain touches privileged registers,
related functions were moved to bringup file.

Change-Id: I1acb66bf84c7c3d9e46d535ea02ca0773bf29735
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>