[SW-120875] gaudi3: remove irrelevant PDMA comment
The comment stated that we need to verify that the PDMA channel
is idle before attempting to change its related configurations.
However, PDMA code in LKD is written such that we simply can't change
its configurations while it isn't idle.
If we accidently do that, then it means we will have to hit a bug.

Change-Id: I3d093ebc70520bcd67aa7c2e483e704484fa6d1d
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>