[SW-119625] greco: route DEC4 to DEC9 if DEC binning in DCORE0
Currently driver follows F/W info about decoder binning, and bin out a
decoder either in DCORE0 or DCORE1.
However when binning is done in DCORE0, F/W still expects to have 9
consecutive decoders, and it ends up with bad F/W accesses to DEC4 in
DCORE0.

To have these 9 consecutive decoders in this case, add a routing
configuration such that accesses to the binned-out DEC4 will be routed
to the non-faulty DEC9.

3 LBW_DECODE entries in the RTR are required for this configuration,
because a single DEC CFG space is 24KB and it needs to be split into
3 x 8KB entries.

Change-Id: Id166d5f8304017e576bb2b4d13129dad1801300a
Signed-off-by: Tomer Tayar <ttayar@habana.ai>