[SW-112986] habanalabs: optimize timestamp registration handling
As part of handling the ts reg node in the user interrupts handler,
we build a list of nodes each contain info about TS/CQ buff to put
refcount to in the WQ context (not the irq context).
Currently those nodes are allocated dynamically in the interrupt
handling context, this patch will change this by adding extra memory
to the ts node data structure to hold this node.

Change-Id: I13be7763898b2307552e4bf0e35c549a7e8de5a8
Signed-off-by: farah kassabri <fkassabri@habana.ai>