[SW-122875] gaudi3: fix bug in early init flow
As part of this flow driver reads preboot status and capability.
if it fails to do that then it might asks for hard reset.
the problem is that early init happens before sw_init
where device asic_specific structure is allocated.
this pointer is used in the hw_fini function to identify whether
the psoc will handle the reset or not, and in such case driver will
crash.

Change-Id: I33aa1edadeff4b778e6eec565ec8e66ed8ffcd6a
Signed-off-by: farah kassabri <fkassabri@habana.ai>