[SW-112961] nic: handle continuous port error events
Add code which tracks every port reset which is triggered due to an error
event.
In case 3 sequential port resets were triggered due to the same syndrom
inside a window time frame of 15 seconds, there is probably a basic issue
with this port configuration.
Therefore, in order to avoid infinite port resets, this port will be
disabled and future H/W initializations of it will be failed.

Change-Id: Ib0f476238f5c6a7d6b8d9d854590dee0ededaad7
Signed-off-by: Sagiv Ozeri <sozeri@habana.ai>