[SW-123213] habanalabs/gaudi2/gaudi3: add a new HL_CS_FLAGS_ENGINES_COMMAND
This addition was added to support a new design of the EDP power test.
The EDP test is used to verify the functionality of the ASIC's power supply
by generating a fast power usage transient toggling (low <--> high).
It will be used to STALL/RESUME TPCs/MMEs/EDMAs, and/or HALT/RUN
engine-ARCs (known as engine-cores) on demand.
Having said that, this new command will be a part of the CS_IOCTL,
and will eventually replace HL_CS_FLAGS_ENGINE_CORE_COMMAND that will
become obsolete.

Change-Id: Icabeab4981e8e11d09c9f08604e9b87e3b63f22d
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>