[SW-0] habanalabs: use scnprintf() in print_device_in_use_info()
compose_device_in_use_info() was added to handle the snprintf() return
value in a single place.
However, the buffer size in print_device_in_use_info() is set such that
it would be enough for the max possible print, so
compose_device_in_use_info() is not really needed.
Moreover, scnprintf() can be used instead of snprintf(), to save the
check if the return value larger than the given size.

Change-Id: Ic5d4c2edf80fceacae83539f3751496af6ee67ab
Signed-off-by: Tomer Tayar <ttayar@habana.ai>