[SW-114047] accel/habanalabs: register compute device as an accel device
Register the compute device as an accel device, and remove the creation
of the habanalabs compute char device.
The habanalabs class is still required for the control device.

IOCTL in this patch are still handled by the current driver handler.
Moving to DRM IOCTL handling requires moving the IOCTLs numbers to a
specific range, so it will be handled in subsequent patches.

Change-Id: I20c2cba2d792fb3b1890932b1dc068c318d33b31
Signed-off-by: Tomer Tayar <ttayar@habana.ai>