[SW-123887] simulator: fix device id value in release function
Each device consumes 2 minors, for main device and for control device.
That's why when we set the simulator device id in the device name
in simulator_start routine we divide the id by 2,
since the minor that we'll get form idr will be always multiple of 2.
In simulator_release routine we print the device id that is being
released and there we should do the same and divide by 2
in order to print the correct main device id.

Change-Id: I37a0623d6a98d0b2150da4fb243d157952aa969e
Signed-off-by: farah kassabri <fkassabri@habana.ai>