[SW-122718] habanalabs/gaudi2: move soft-reset wait to soft-reset execute
We plan to do soft-reset either by r/w register or using cpucp packet
depending on the FW version. We don't want to check FW version in two
different places for that (execute soft-reset and wait to soft-reset)
so move the waiting to gaudi2_execute_soft_reset. This also makes sense
since the cpucp also does the waiting.

Change-Id: I296225feade5872fa279442ca93a024f5a2be2ff
Signed-off-by: Dafna Hirschfeld <dhirschfeld@habana.ai>