[SW-120875] gaudi3: use the correct register to set/get device HW state
So far, we didn't use the correct register to get HW status.
Now that it has been fixed, and the HW status reg points to
the right register we can get the correct device HW state.
As well, for an accurate read, we must set this register on boot-time,
allowing this register to be cleared by the HW upon reset.

Change-Id: I2b610787f00abb9558a34e3850a12385c5861368
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>