[SW-123888] habanalabs/gaudi3/sim: do not reset if CQ-completion job fails
A job execution with CQ-completion mechanism is used by several
engines in LKD (currently, CBC/H/MMU invalidation, PDMA transaction).
So far, upon a job failure (i.e., CQ-completion timed-out), LKD used
to trigger a reset-flow, independent of the initiator that failed.
Scheduling a reset when CQ-completion timed-out, prevents from the caller-
itself the ability to self-handle the issue (e.g., user is killed on
hard-reset), or even knowing what went wrong.
Having said that, we can't do soft-reset while a user is holding the
device, and a manual hard-reset is an unwanted & extreme measure.
Therefore, CQ-completion failures will not trigger a hard-reset anymore,
such that the caller will be the one who decides of the required action
upon a failure.
If any of the above initiators fail during boot-time, we simply fail the
driver-load.
If any of them fail in steady-state, the caller can choose to schedule
a reset.

Change-Id: Iec4b7db72da39cd223fd6b08403b8d4cecd7ffa0
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>