[SW-122762] debugfs: add debugfs to override link status
In Gaudi3 while working on PLDM we encounter a situation,
where we need to work in non MAC loopback mode. In this scenario
we might encounter a situation where speed bridge doesn't generate
an UP link event to the driver. This affects the external ports
as they will be considered in down mode. Under such state
we are not allowed by the rdma-core to change state of the QP
and hence the test would fail.
To overcome this and not to mask out real issues we should
override the link status prior preforming this test.

Change-Id: I51f085be8dca163e129725890b3ce8809a92af7c
Signed-off-by: Andrey Agranovich <aagranovich@habana.ai>