[SW-0] nic: fix collective QPs destroy flow
When allocating a collective QP, the driver allocates a new collective
ID and also allocates a new QP per port with this ID.
Later, the user might set only the QPs of some of the ports.
Finally, when the user destroys these specific QPs, the driver needs to
take care of destroying also the unset QPs in the rest of the ports and
finally to remove the collective ID from the IDR.
The way it does that is to call tha generic destroy flow for all other
unset QPs.

In case that the user asked to allocate a new collective QP right after
the collective ID was destroyed, the same ID will be allocated and then
an old unset QP that is being destroyed now could potentially find and
destroy it since there is no set QPs under this collective QP.

In order to avoid such a case and take care of all potential races,
the last set QP will take care of destroying all other QPs and the
last QP that is being destroyed will take care of destroying the
collective QP and remove its ID from the idr.

In addition:
1. QP will consider as set when changing its state from RESET to INIT.
2. Move qp_register to be done in alloc_coll_qp like it is done for
   generic QPs.

Change-Id: Ifbf946655f5e3ff326f36ea853de2308e955ec1e
Signed-off-by: Sagiv Ozeri <sozeri@habana.ai>