[SW-123246] hl_sni: NIC dedicated driver
Move all NIC infrastructure code to a new dedicated driver -
habanalabs_sni ("Shared Network Interface").
Now that we are part of the accel subsystem, we cannot have the NIC code
embedded in the compute driver - we need to move it to the net subsystem.
The SNI driver loads the ETH and IB devices.

Change-Id: I769da0987d124ae6328bce0fa70aa4aaabbe35a2
Signed-off-by: Omer Shpigelman <oshpigelman@habana.ai>