[SW-112986] accel/habanalabs: optimize timestamp registration handler
Currently we use dynamic allocation inside the irq handler
in order to allocate free node to be used for the free jobs.
this operation is expensive specially when we deal with large
burst of events records that get released at the same time.
the alternative is to have pre allocated pool of free nodes
and just fetch nodes from this pool at irq handling time instead
of allocating them.

As part of the optimization also update the unregister flow
upon re-using a timestamp record, by making the operation much
simpler and quicker. since this function in the past used for implementing
unregister timestamp record ioctl, which was deleted now,
so we already have the record in the registration
flow and now we just seek to re-use with different interrupt then no need
to look for buffer according to the user handle.

Change-Id: Ibdc318a8f901a147036f10f8942a32b7df7fa23a
Signed-off-by: farah kassabri <fkassabri@habana.ai>