[SW-112986] habanalabs: optimize timestamp registration handler
Currently we use dynamic allocation inside the irq handler
in order to allocate free node to be used for the free jobs.
this operation is expensive specially when we deal with large
burst of events records that get released at the same time.
the alternative is to have pre allocated pool of free nodes
and just fetch nodes from this pool at irq handling time instead
of allocating them.

Change-Id: Idecee9eb192794e537946066652a960deac53eac
Signed-off-by: farah kassabri <fkassabri@habana.ai>