[SW-136270] gaudi3: avoid interrupt storm
In case an empty eqe is received, in many cases it will cause an
interrupt storm because the root cause is not getting cleared.
The interrupt storm is a symptom and not the root cause of the bug,
but it need to be avoided.

Fixed by avoiding interrupt unmask in case its cause is empty.

Change-Id: I24f1f5974642a5bcc8238d5b6cb4c6414744e473
Signed-off-by: Dani Liberman <dliberman@habana.ai>