[SW-136181] accel/habanalabs: update timestamp registration handling
In case of handling re-use the same ts node for different interrupt,
the driver is not protected since it is using the lock of the new
interrupt, and not the lock of the interrupt it's already registered on.
We need timestamp registration to handle re-use of the same node,
by calling unregister timestamp function protected by the lock of the
interrupt it's currently registered on, then do the new
registration.
This patch will update the timestamp registration driver to store the
interrupt data during the registration and will handle the re-use as
mentioned above, in addition to some refactor.

Change-Id: Icb706b3fccf50e0ec4eb7bd172c6b18fc789b052
Signed-off-by: farah kassabri <fkassabri@habana.ai>