[SW-0] compat: prevent grep error messages when hmm.h is not found
Makefile.compat has several checks in hmm.h, and some of them require
several grep commands in a pipe.
The redirection of error messages there is done only for the last grep,
so when hmm.h is not part of some kernel version, a "No such file or
directory" message will be printed.
To prevent error messages for all grep commands, put the bash expression
inside parentheses.

Change-Id: I1ad5571e9743b9b5e4e5ca1ec5b4a6c0518ad1e3
Signed-off-by: Tomer Tayar <ttayar@habana.ai>