[SW-120875] gaudi3: move relevant page_fault_queue code to sw_init
This is done to fix a bug where in case device-reset is executed by FW,
we skip memory deallocation of the page fault queue, afterwhich a new
memory allocation will be done, running over the previous one.
As well, there's no reason to deallocate/reallocate memory every time
we reset the device, that's exactly what sw_init functionality is for.

Change-Id: I83c3322d1984248c52fa8cb471a523b657c8d91e
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>