[SW-114047] gaudi: remove HBM debugfs code
In a future change the device is going to be registered as an accel
device. As such, the debugfs root folder will be created by the accel
subsystem, and it will be done only after registering the device and
exposing the interfaces to user.
However, the gaudi HBM debugfs files are created earlier as part of
hw_init.

While it is possible to add an infrastructure that replaces the direct
debugfs functions calls with saving the info of these debugfs files and
postponing their creation, it doesn't seem worth the trouble because
these HBM debugfs files are created only if the driver performs the HBM
initialization, which is not a valid option anymore on ASIC.
Instead, simply remove the whole code for these debugfs files.

Change-Id: I80e9cd4afac3916eaf257974c4e2579c7eff6a67
Signed-off-by: Tomer Tayar <ttayar@habana.ai>