[SW-114047] accel/habanalabs: expose debugfs files with cdev and sysfs files
Currently the debugfs root folder and files for a device are created at
an early step, before the device initialization and before the char
device and sysfs files are exposed to user.
As there is no real reason not to do it together with the device
creation, postpone it to be done right afterwards.

The initialization of the debugfs entry structure is left in its current
position, because it is used before creating the files.

Change-Id: Iedbc87d755b422e31815a6e5acf1365034a3226a
Signed-off-by: Tomer Tayar <ttayar@habana.ai>