[SW-123877] accel/habanalabs: sync EQ interrupt in hard reset
Receiving EQs from FW, while the device is in hard reset, causes
a warning message in Driver log. The message may point on a
problem in the Driver or FW. But It also can appear as a result
of EQs that have been sent from FW, just before the hard reset.
In order to avoid receiving EQs from FW, while the device is reset
and is already in 'disabled' mode, a sync EQ interrupt is added, right
before setting the device to 'disabled'.

Change-Id: I6c6328f4cbc5c98579bd57a23f9846229aedb6f1
Signed-off-by: Tal Cohen <talcohen@habana.ai>