[SW-114047] simulator: add accel device registration
Add support for the accel device registration when working with a
simulator.

The registration is done using a new platform device that acts as a
parent device.
The DRM initialization requires such a parent device, which in case of
simulator could have been the simulator char device.
This char device has no unregister_driver and driver detach that will
lead to releasing devrese entries, including a required call to
drm_dev_put().
A platform device however does have this required unregistration.

Similarly to IDR allocation for ASIC, also for simulator allocate only a
single ID.

Change-Id: Icb371f784811a9cea068620c06ba42a4e1228ba3
Signed-off-by: Tomer Tayar <ttayar@habana.ai>