[SW-126413] nic: clear RAS counters on port initialization
We should clear RAS counters when link is brought up, either
internal or external port. RAS counters are derived partly from
ethtool counters, which exist for any port. Hence we should clear them
too.
We rely on already existing flow which was not utilized until now.
We consider ports as fresh initialized when we open a port.
Hence counters should be cleared.

Change-Id: I7963438d0a5fcda546a41352844959beb27a0e1b
Signed-off-by: Andrey Agranovich <aagranovich@habana.ai>