[SW-136530] gaudi3: allow a binned ASIC to boot even with disabled ROT/DEC
Now that binning is defaultly done by the Preboot, LKD is trying
to validate the binning itself.
However, a strict ancient rule requests all engines to be enabled
before validating binning, otherwise driver load fails.
This causes a problem when running PLDM images, where not all
engines are necessarily available.
Therefore, this refactor allows the driver to boot even if not all
engines are enabled.
The disabled engines will be taken into account along with the eFuse's
binned info, altogether.

Change-Id: I95b4363b049e5047be24463210d47ff4bab89a90
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>