[SW-0] gaudi3: add function to validate expected eqe data size
In various events handlers there is a repeating pattern of checking the
actual data size vs the expected data size.
To save copy-paste mistakes when it is added to new handlers, and to
make it easier to do changes in this check in the future, e.g. to modify
the error message, move it into a dedicated function.

Change-Id: Ifc26688bcc2f0f724b110b0cd6106eec78c02d9d
Signed-off-by: Tomer Tayar <ttayar@habana.ai>