[SW-127124] gaudi2: enable CQ overrun during QP invalidation
Allow CQ overrun to make sure QP drain is successful. In case PFC is sent
due to CQ overflow, no Tx can be sent until CQ releases the back-pressure.
If in the meanwhile a QP needs to be invalidated, no tx will be sent in
the QP drain stage. This will cause Tx slices to be stuck after the QP
drain stage has finished. Next when the CQ will be destroyed, it will
release the back-pressure, causing the stuck Tx slices to be sent (as
there's no more back-pressure). Since no QP is allocated anymore, AXI
errors and QP invalid errors will be received. As a workaround to the
issue above, allow overrun to the assoicated CQ of the invalidated QP.
This will release the back-pressure before the drain stage, and will
allow all needed tx packets to be drained successfully. Once the drain
stage is done, and the QP is cleared, disable CQ overrun.

Change-Id: I2084d6fcc6b9eba4d71cd4d229d26a87c85f1f81
Signed-off-by: David Meriin <dmeriin@habana.ai>