[SW-138356] accel/habanalabs: poll for CPU status update following WFE cmd
Currently, we rely on COMMS protocol's ack to verify that WFE command
has been acknowledged by the FW, however, this does not promise that
the CPU status has been updated respectively.
Although unlikely, this could trigger a race since LKD expects the
CPU to be halted at that stage, but it might not be.
Therefore, we increase WFE's robustness by polling on the PSOC
GLOBAL_CONF register that will eventually be updated.

Change-Id: Ibe709cfd94011de79156225be156e823196b80a2
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>