[SW-120212] accel/habanalabs: avoid recognize psoc razwi when address is zero
In Gaudi2 asic, PSOC RAZWI may cause in HBW or LBW. The address that
caused the error is read from HW register and printed by the Driver.
There are cases where the Driver receives an indication on PSOC
RAZWI error but the address value is zero.
The Driver should not "count" a PSOC RAZWI event error when the
caused the address is zeroed.

Change-Id: I95e4140f81338dcd3c92c09c6de178b529bfc493
Signed-off-by: Tal Cohen <talcohen@habana.ai>