[SW-137001] accel/habanalabs: fix crash in gaudi3 fpga by setting supports_cb_mapping if pmmu enabled
it is not possible to map cb to device va space if the pmmu is disabled.
So set 'supports_cb_mapping' to true only if pmmu is enabled.
The function hl_cb_va_pool_init checks only for 'supports_cb_mapping'
and not for 'mmu_enable' causing a crash in gaudi3 fpga where mmu is
disabled.

Change-Id: I9574f3284a2dbb981e5166828eb006cb233ed535
Signed-off-by: Dafna Hirschfeld <dhirschfeld@habana.ai>