[SW-0] gaudi2 phy: port 9 toggling WA
Port 8 is an external port which will usually be brought UP after all
the internal ports are UP. Due to NIC macro clock nest dependency, when
PHY reset is called for port 8, port 9 (which is internal) is being
toggled and might lost stabilization.

A W/A to overcome this issue is to perform PHY reset to port 9 right
after, and force its PHY to be reconfigured from start.

Change-Id: Ifb602438fe29aaeb9ba501360bef0a1ad86dbff3
Signed-off-by: Sagiv Ozeri <sozeri@habana.ai>