[SW-0] accel/habanalabs: remove memset of error info when open fd
In some asics, we may not be able to read captured error data before
reset. In this case, when user opens the fd (after previous fd was
closed due to user process termination), the captured error data will
be empty.

Fixed it by removing the memset and resetting only the write enable
flags.

Change-Id: Ia8829c1b068786544ff10e678b894c2bbbe7f839
Signed-off-by: Dani Liberman <dliberman@habana.ai>