[SW-0] simulator: set crashed flag before the corresponding print
Sometimes after a simulator gets closed and there is "Simulator was
closed" message, there is another error message for some operation which
should have been expected to be skipped due to the simulator crashed
flag.
As this can happen due to an interleaved operation between the "was
closed" print and setting the flag, modify the order to first handle the
flag and only then print.

Change-Id: Id8b297e1598196e1952286279042b67b87faee0d
Signed-off-by: Tomer Tayar <ttayar@habana.ai>