[SW-114047] simulator: small code refactor of sim_start/stop functions
Simulator files have both simulator_start/stop and sim_start/stop
functions, which can lead to a confusion.
As sim_start/stop are more internal and responsible for the device
itself, add a _device suffix to these functions to make them more
distinct.

In addition, modify the sim_start functions to receive only the
simulator device structure as a parameter, because other parameters can
be derived from it.

Change-Id: I8f09aa159daa192f1047325c72a3792dc2fdd3e3
Signed-off-by: Tomer Tayar <ttayar@habana.ai>