[SW-123801] accel/habanalabs: don't fail nic reset req when removing device
NIC can ask for the device to hard reset through the debug file system.
However, waiting for this reset flow to finish successfully is possible
only if we're not in a device removal reset flow (rmmod).
If someone decides to remove the device, then the driver waits for
any ongoing reset to finish, or alterantively, flushes any pending
reset requests, after which device_fini routine will be called.
In worst case scenario, we could end waiting long time before device
is removed, which could be mistakenly thought to be a failure.

Change-Id: I4e63934168c3dcfc71ee610a17b6e6bc581da55f
Signed-off-by: Koby Elbaz <kelbaz@habana.ai>