[SW-114047] accel/habanalabs/compat: add debugfs symlinks for accel
Upstream kernel is going to have a change that modifies the debugfs
directory to be named as the device-name instead of as the minor-id.
However, we still need to support older kernel versions that have the
accel subsystem and still have the previous minor-id debugfs
directories.
In addition, the "fake" accel directories when not working with the
accel subsystem, still have the minor-id as the directory name.

To support all versions but still allowing a single method of accessing
only a device-name named directory, add debugfs symlinks when needed:

1. For kernel versions < 6.2 ("fake" accel):
   Replace the current minor-id named directory with a device-name named
   directory, and add a temporary minor-id named symlink to it.
   This symlink should be removed when all move to use only the actual
   debugfs directory.

2. For kernel versions >= 6.2 until the kernel to which this name change
   is added:
   Add a device-name named symlink that points to the minor-id named
   debugfs directory which is created by the accel subsystem.

3. For kernel versions with this name change:
   No symlink is required, as the accel subsystem already creates a
   device-name named directory.

Change-Id: I4e7ba44ac5081f8b1e05ca8ff9dd49ea8793a7c8
Signed-off-by: Tomer Tayar <ttayar@habana.ai>