[SW-114047] accel/habanalabs: create accel control device using accel subsystem class
When registering to the accel subsystem, use its class for the creation
of the accel control device.

This change is separated from the accel subsystem registration change,
because a "manual" creation of a non-accel device using the accel class
won't be accepted in the community and therefore won't be upstreamed.

***** THIS CHANGE SHOULDN'T BE UPSTREAMED *****

Change-Id: I5d0706a60eb5db5aa392eadbbcc20cc3a01b61da
Signed-off-by: Tomer Tayar <ttayar@habana.ai>